pyproject.toml
[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | env
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.poetry]
name = "dash-dep"
version = "0.1.0"
description = ""
authors = ["Team Python"]
readme = "README.md"
packages = [{include = "../dash"}]

[tool.poetry.dependencies]
python = "^3.12"
ansi2html = "1.9.1"
apscheduler = "3.10.4"
async-timeout = "4.0.3"
clickhouse-driver = "0.2.7"
coverage = "7.4.4"
dash-ace = "0.2.1"
dash-bootstrap-components = "1.5.0"
dash-leaflet = "1.0.15"
dash-pivottable = "0.0.2"
databricks-cli = "0.18.0"
deprecated = "1.2.14"
elasticsearch = "8.12.1"
exceptiongroup = "1.2.0"
flask-caching = "2.1.0"
geobuf = "1.1.1"
ipython = "8.22.2"
json-logging = "1.3.0"
missingno = "0.5.2"
mlflow = "2.11.1"
natsort = "8.4.0"
nbconvert = "7.16.2"
nltk = "3.8.1"
openpyxl = "3.1.2"
prometheus-flask-exporter = "0.23.0"
psycopg2-binary = "2.9.9"
pydantic = "1.10.14"
pymorphy2-dicts-ru = "2.4.417127.4579844"
pyodbc = "5.1.0"
pytz-deprecation-shim = "0.1.0.post0"
redis = "5.0.3"
sentry-sdk = "1.42.0"
tangled-up-in-unicode = "0.2.0"
toml = "0.10.2"
tomli = "2.0.1"
urlextract = "1.9.0"
wheel = "0.43.0"
whoosh = "2.7.4"
xlrd = "2.0.1"
xlsxwriter = "3.2.0"
yarg = "0.1.9"
pymorphy2 = { git = "https://pymorphy:VKmc8ZHESCu-a4vqkis5@gitlab.sima-land.ru/dev-dep/python/pymorphy2.git", tag = "v0.0.1" }
statsmodels = "0.14.1"
aiohttp = "3.9.3"
grpcio = "1.62.1"
grpcio-tools = "1.62.1"
celery = "5.4.0"
flask = "3.0.3"
gql = {extras = ["all"], version = "3.5.0"}
python-gitlab = "5.0.0"
pytest = "8.1.1"
minio = "^7.2.12"
boto3 = "^1.35.78"

[tool.poetry.group.dev.dependencies]
black = "24.3.0"
flake8 = "7.0.0"
pre-commit = "3.6.2"
pytest = "8.1.1"
mock = "5.1.0"
sqlparse = "0.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
